+++ include/kernel/arch/x86_64/interrupts.hpp
@@ -9,54 +9,36 @@ namespace arch {
 namespace x86_64 {
 
 /**
- * @brief CPU Register Stack Frame saved during Interrupt Service Routine (ISR) - 32-bit
+ * @brief Registers pushed on stack during interrupt entry
  */
-struct __attribute__((packed)) InterruptRegisters {
-    uint32_t edi;             ///< Saved EDI from pusha
-    uint32_t esi;             ///< Saved ESI from pusha
-    uint32_t ebp;             ///< Saved EBP from pusha
-    uint32_t esp;             ///< Saved ESP from pusha
-    uint32_t ebx;             ///< Saved EBX from pusha
-    uint32_t edx;             ///< Saved EDX from pusha
-    uint32_t ecx;             ///< Saved ECX from pusha
-    uint32_t eax;             ///< Saved EAX from pusha
-
-    uint32_t int_no;          ///< Interrupt Vector number (0-255)
-    uint32_t err_code;        ///< Error Code (dummy or hardware)
-
-    uint32_t eip;             ///< Instruction Pointer pushed by hardware
-    uint32_t cs;              ///< Code Segment pushed by hardware
-    uint32_t eflags;          ///< EFLAGS register pushed by hardware
+struct __attribute__((packed)) registers_t {
+    uint32_t ds;                                     ///< Data segment selector
+    uint32_t edi, esi, ebp, esp, ebx, edx, ecx, eax; ///< Pushed by pusha
+    uint32_t int_no, err_code;                       ///< Interrupt number and error code
+    uint32_t eip, cs, eflags, useresp, ss;           ///< Pushed by CPU automatically
 };
 
-typedef void (*ISRHandler)(InterruptRegisters* regs);
+typedef void (*isr_t)(registers_t* regs);
 
 /**
- * @brief Interrupt Manager Class
+ * @brief Register callback handler for interrupt vector
+ * @param n Vector index (0-255)
+ * @param handler Callback function pointer
  */
-class Interrupts {
-private:
-    static ISRHandler m_handlers[256];
+void register_interrupt_handler(uint8_t n, isr_t handler);
 
-public:
-    /**
-     * @brief Register a custom C++ handler function for an interrupt vector
-     * @param vector Vector index (0-255)
-     * @param handler Function pointer to handler
-     */
-    static void register_handler(uint8_t vector, ISRHandler handler);
-
-    /**
-     * @brief C++ main interrupt dispatcher invoked by assembly ISR stubs
-     * @param regs Pointer to register stack frame
-     */
-    static void dispatch(InterruptRegisters* regs);
-};
+/**
+ * @brief Initialize ISR exception stubs, IRQ stubs, and IDT gates
+ */
+void interrupts_init();
 
 } // namespace x86_64
 } // namespace arch
 } // namespace nebula
 
-extern "C" void cpp_interrupt_handler(nebula::arch::x86_64::InterruptRegisters* regs);
+extern "C" {
+    void isr_handler(nebula::arch::x86_64::registers_t* regs);
+    void irq_handler(nebula::arch::x86_64::registers_t* regs);
+}
 
 #endif // NEBULA_ARCH_X86_64_INTERRUPTS_HPP