+++ kernel/arch/x86_64/interrupts/interrupts.cpp
@@ -1,78 +1,96 @@
 /**
  * @file interrupts.cpp
- * @brief C++ Interrupt Dispatcher implementation for x86_64
+ * @brief Interrupt service routine dispatchers and registration for Nebula OS
  * @author Nebula OS Team
  */
 
 #include "../../../../include/kernel/arch/x86_64/interrupts.hpp"
+#include "../../../../include/kernel/arch/x86_64/idt.hpp"
 #include "../../../../include/kernel/arch/x86_64/pic.hpp"
-#include "../../../../include/kernel/drivers/console.hpp"
+#include "../../../../include/kernel/drivers/serial.hpp"
 
-using nebula::drivers::vga_write_at;
-using nebula::drivers::COLOR_WHITE;
-using nebula::drivers::COLOR_RED;
-using nebula::drivers::COLOR_YELLOW;
+extern "C" {
+    // 32 CPU Exception Stubs
+    void isr0();  void isr1();  void isr2();  void isr3();
+    void isr4();  void isr5();  void isr6();  void isr7();
+    void isr8();  void isr9();  void isr10(); void isr11();
+    void isr12(); void isr13(); void isr14(); void isr15();
+    void isr16(); void isr17(); void isr18(); void isr19();
+    void isr20(); void isr21(); void isr22(); void isr23();
+    void isr24(); void isr25(); void isr26(); void isr27();
+    void isr28(); void isr29(); void isr30(); void isr31();
+
+    // 16 IRQ Hardware Stubs
+    void irq0();  void irq1();  void irq2();  void irq3();
+    void irq4();  void irq5();  void irq6();  void irq7();
+    void irq8();  void irq9();  void irq10(); void irq11();
+    void irq12(); void irq13(); void irq14(); void irq15();
+}
 
 namespace nebula {
 namespace arch {
 namespace x86_64 {
 
-ISRHandler Interrupts::m_handlers[256] = { nullptr };
+typedef void (*isr_stub_t)();
 
-static const char* exception_messages[32] = {
-    "Division By Zero",
-    "Debug",
-    "Non-Maskable Interrupt",
-    "Breakpoint",
-    "Overflow",
-    "BOUND Range Exceeded",
-    "Invalid Opcode",
-    "Device Not Available",
-    "Double Fault",
-    "Coprocessor Segment Overrun",
-    "Invalid TSS",
-    "Segment Not Present",
-    "Stack-Segment Fault",
-    "General Protection Fault",
-    "Page Fault",
-    "Reserved Exception 15",
-    "x87 FPU Floating-Point Error",
-    "Alignment Check",
-    "Machine Check",
-    "SIMD Floating-Point Exception",
-    "Virtualization Exception",
-    "Control Protection Exception",
-    "Reserved Exception 22",
-    "Reserved Exception 23",
-    "Reserved Exception 24",
-    "Reserved Exception 25",
-    "Reserved Exception 26",
-    "Reserved Exception 27",
-    "Hypervisor Injection Exception",
-    "VMM Communication Exception",
-    "Security Exception",
-    "Reserved Exception 31"
+// Array of 32 CPU Exception stub function pointers
+static const isr_stub_t isr_stubs[32] = {
+    isr0,  isr1,  isr2,  isr3,  isr4,  isr5,  isr6,  isr7,
+    isr8,  isr9,  isr10, isr11, isr12, isr13, isr14, isr15,
+    isr16, isr17, isr18, isr19, isr20, isr21, isr22, isr23,
+    isr24, isr25, isr26, isr27, isr28, isr29, isr30, isr31
 };
 
-void Interrupts::register_handler(uint8_t vector, ISRHandler handler) {
-    m_handlers[vector] = handler;
+// Array of 16 IRQ Hardware stub function pointers
+static const isr_stub_t irq_stubs[16] = {
+    irq0,  irq1,  irq2,  irq3,  irq4,  irq5,  irq6,  irq7,
+    irq8,  irq9,  irq10, irq11, irq12, irq13, irq14, irq15
+};
+
+static isr_t interrupt_handlers[256] = { nullptr };
+
+static const char* exception_messages[] = {
+    "Division By Zero",
+    "Debug",
+    "Non Maskable Interrupt",
+    "Breakpoint",
+    "Into Detected Overflow",
+    "Out of Bounds",
+    "Invalid Opcode",
+    "No Coprocessor",
+    "Double Fault",
+    "Coprocessor Segment Overrun",
+    "Bad TSS",
+    "Segment Not Present",
+    "Stack Fault",
+    "General Protection Fault (#GP)",
+    "Page Fault (#PF)",
+    "Unknown Interrupt",
+    "Coprocessor Fault",
+    "Alignment Check",
+    "Machine Check",
+    "SIMD Floating Point",
+    "Virtualization Exception",
+    "Control Protection Exception",
+    "Reserved", "Reserved", "Reserved", "Reserved",
+    "Reserved", "Reserved", "Reserved", "Reserved",
+    "Security Exception",
+    "Reserved"
+};
+
+void register_interrupt_handler(uint8_t n, isr_t handler) {
+    interrupt_handlers[n] = handler;
 }
 
-void Interrupts::dispatch(InterruptRegisters* regs) {
-    if (m_handlers[regs->int_no] != nullptr) {
-        m_handlers[regs->int_no](regs);
-    } else if (regs->int_no < 32) {
-        // Unhandled CPU Exception
-        vga_write_at(0, 4, "=== CPU EXCEPTION ===", COLOR_WHITE, COLOR_RED);
-        vga_write_at(0, 5, exception_messages[regs->int_no], COLOR_YELLOW, COLOR_RED);
-        while (true) {
-            asm volatile ("cli; hlt");
-        }
+void interrupts_init() {
+    // Register 32 CPU Exception gates (Vectors 0-31)
+    for (uint8_t i = 0; i < 32; i++) {
+        IDT::set_gate(i, reinterpret_cast<uint32_t>(isr_stubs[i]), 0x08, 0x8E);
     }
 
-    // Send EOI for hardware IRQs (vectors 32 - 47)
-    if (regs->int_no >= 32 && regs->int_no <= 47) {
-        PIC::send_eoi(static_cast<uint8_t>(regs->int_no - 32));
+    // Register 16 IRQ Hardware gates (Vectors 32-47)
+    for (uint8_t i = 0; i < 16; i++) {
+        IDT::set_gate(32 + i, reinterpret_cast<uint32_t>(irq_stubs[i]), 0x08, 0x8E);
     }
 }
 
@@ -80,6 +98,37 @@ void Interrupts::dispatch(InterruptRegisters* regs) {
 } // namespace arch
 } // namespace nebula
 
-extern "C" void cpp_interrupt_handler(nebula::arch::x86_64::InterruptRegisters* regs) {
-    nebula::arch::x86_64::Interrupts::dispatch(regs);
+extern "C" {
+
+void isr_handler(nebula::arch::x86_64::registers_t* regs) {
+    if (!regs) return;
+
+    uint8_t int_no = static_cast<uint8_t>(regs->int_no);
+
+    if (nebula::arch::x86_64::interrupt_handlers[int_no] != nullptr) {
+        nebula::arch::x86_64::interrupt_handlers[int_no](regs);
+    } else {
+        nebula::drivers::Serial::write_string("[CPU EXCEPTION] ");
+        if (int_no < 32) {
+            nebula::drivers::Serial::write_string(nebula::arch::x86_64::exception_messages[int_no]);
+        } else {
+            nebula::drivers::Serial::write_string("Unhandled Exception");
+        }
+        nebula::drivers::Serial::write_string("\n");
+    }
+}
+
+void irq_handler(nebula::arch::x86_64::registers_t* regs) {
+    if (!regs) return;
+
+    uint8_t int_no = static_cast<uint8_t>(regs->int_no);
+    uint8_t irq_no = int_no - 32;
+
+    if (nebula::arch::x86_64::interrupt_handlers[int_no] != nullptr) {
+        nebula::arch::x86_64::interrupt_handlers[int_no](regs);
+    }
+
+    nebula::arch::x86_64::PIC::send_eoi(irq_no);
+}
+
 }